refactoring runmat-core crate for maintainability
