+++ crates/runmat-core/src/fusion/mod.rs
@@ -1,0 +1,5 @@
+mod snapshot;
+mod types;
+
+pub(crate) use snapshot::build_fusion_snapshot;
+pub use types::*;
