+++ crates/runmat-core/src/workspace/mod.rs
@@ -1,0 +1,14 @@
+mod emit;
+mod materialize;
+mod types;
+
+pub(crate) use emit::{
+    determine_display_label_from_context, format_type_info,
+    last_displayable_statement_emit_disposition, last_emit_var_index, last_expr_emits_value,
+    last_unsuppressed_assign_var, workspace_entry, FinalStmtEmitDisposition,
+};
+pub(crate) use materialize::{
+    gather_gpu_preview_values, gpu_dtype_label, gpu_size_bytes, slice_value_for_preview,
+    MATERIALIZE_DEFAULT_LIMIT,
+};
+pub use types::*;
